pyproject.toml
[tool.poetry]
name = "djangodockerproj"
version = "0.1.0"
description = ""
authors = ["AlexanderZug <alexandertsygankov.unterwegs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
celery = "^5.3.1"
redis = "^4.6.0"
flower = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
